+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 1. Avoid memory allocation when print help
 2. Avoid memory allocation for pattern matching
+3. Able to shuffle parameterized tests
 
 ### Fixed
 1. Fix: pattern not working on parameterized tests